export directives to commons | resolve changes @blat @crashillo
