+++ app/javascript/gobierto_data/lib/commons.js
@@ -1,1 +1,45 @@
 export const baseUrl = `${location.origin}/api/v1/data`
+
+export const CommonsMixin = {
+  directives: {
+    clickoutside: {
+      bind: function(el, binding, vnode) {
+        el.clickOutsideEvent = function(event) {
+          if (!(el == event.target || el.contains(event.target))) {
+            vnode.context[binding.expression](event);
+          }
+        };
+        document.body.addEventListener('click', el.clickOutsideEvent)
+      },
+      unbind: function(el) {
+        document.body.removeEventListener('click', el.clickOutsideEvent)
+      },
+      stopProp(event) { event.stopPropagation() }
+    },
+    closable : {
+      bind (el, binding, vnode) {
+        let handleOutsideClick
+        handleOutsideClick = (e) => {
+          e.stopPropagation()
+          const { handler, exclude } = binding.value
+          let clickedOnExcludedEl = false
+          exclude.forEach(refName => {
+            if (!clickedOnExcludedEl) {
+              const excludedEl = vnode.context.$refs[refName]
+              clickedOnExcludedEl = excludedEl.contains(e.target)
+            }
+          })
+          if (!el.contains(e.target) && !clickedOnExcludedEl) {
+            vnode.context[handler]()
+          }
+        }
+        document.addEventListener('click', handleOutsideClick)
+        document.addEventListener('touchstart', handleOutsideClick)
+      },
+      unbind () {
+        document.removeEventListener('click', handleOutsideClick)
+        document.removeEventListener('touchstart', handleOutsideClick)
+      }
+    }
+  },
+}
